requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv
httpx
pydantic